edit product, delete confirmations, add to db

+++ components/CatalogueList.js
@@ -11,16 +11,27 @@ import { getDateTime } from "../utils/ProductUtils";
 export const CatalogueList = ({ productList, productSets, getProductSets, setLoading }) => {
     const { theme } = useTheme();
     const [selected, setSelected] = useState(null);
+    const [showDelete, setShowDelete] = useState(false);
+    const [deleteSelect, setDeleteSelect] = useState(null);
     const [show, setShow] = useState(false);
     const { setTab, showToast, SERVER_ADDRESS } = useAppContext();
 
     const onDelete = async (productSetName) => {
+
+        if (!showDelete) {
+            setDeleteSelect(productSetName);
+            setShowDelete(true);
+            return;
+        };
+
         setLoading(true);
 
         await axios.delete(`${SERVER_ADDRESS}/api/productset/${getProductID(productSetName)}`)
         .then((response) => {
             showToast(response.data.message);
             getProductSets();
+            setDeleteSelect(null);
+            setShowDelete(false);
         }).catch((err) => {
             showToast(err.message);
         });
@@ -30,12 +41,56 @@ export const CatalogueList = ({ productList, productSets, getProductSets, setLoa
 
     return (
         <View style={{ width: "100%", flex: 1 }}>
+
+            <Overlay
+                isVisible={showDelete}
+                onBackdropPress={() => {
+                    setShowDelete(false);
+                    setDeleteSelect(null);
+                }}
+                overlayStyle={{ width: "70%", borderRadius: 5 }}>
+                <View style={{ width: "100%", padding: 10 }}>
+                    <Text style={textStyles.header2}>
+                        Delete {deleteSelect}?
+                    </Text>
+                    <View
+                        style={{
+                            flexDirection: "row",
+                            justifyContent: "flex-end",
+                        }}>
+                        <Button
+                            onPress={() => {
+                                setShowDelete(false);
+                                setDeleteSelect(null);
+                            }}
+                            containerStyle={{ margin: 5 }}
+                            type="clear">
+                            <Text style={{ color: theme.colors.Text }}>
+                                Cancel
+                            </Text>
+                        </Button>
+                        <Button
+                            onPress={() => onDelete(selected?.displayName)}
+                            containerStyle={{
+                                margin: 5,
+                                backgroundColor: theme.colors.error,
+                            }}
+                            type="clear">
+                            <Text style={{ color: theme.colors.buttonText }}>
+                                Delete
+                            </Text>
+                        </Button>
+                    </View>
+                </View>
+            </Overlay>
+
             <Overlay
                 isVisible={show}
                 onBackdropPress={() => setShow(false)}
                 overlayStyle={{ width: "90%", borderRadius: 5 }}>
                 <CatalogueProductList productList={productList} productSet={selected} />
             </Overlay>
+
             <ScrollView style={{ width: "100%" }}>
                 {productSets?.map((productSet, i) => {
                     return (
@@ -72,19 +127,6 @@ export const CatalogueList = ({ productList, productSets, getProductSets, setLoa
                                         </Text>
                                     </View>
                                     <View style={{ flexDirection: "row" }}>
-                                        <Button
-                                            onPress={() =>
-                                                showToast(
-                                                    "Edit : not implemented yet"
-                                                )
-                                            }
-                                            containerStyle={{
-                                                margin: 5,
-                                                borderRadius: 50,
-                                            }}
-                                            type="clear">
-                                            <Icon name="edit" />
-                                        </Button>
                                         <Button
                                             onPress={() =>
                                                 onDelete(productSet.displayName)
